python_tasks/requirements.txt
schedule>=1.2.0
psycopg2-binary>=2.9.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
greenlet>=3.0.0
python-dotenv>=1.0.0
markdown>=3.4.0
snowflake-connector-python>=3.0.0
pyyaml>=6.0

pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Industry Standard Data Quality
great_expectations>=0.18.0
